Issue #76. Move handlers.
